requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
sqlalchemy==2.0.32
duckdb==1.2.0
pydantic==2.8.2
python-dotenv==1.0.1
typer==0.12.3
rich==13.7.1
plotly==5.22.0
pyyaml==6.0.1
pytest==8.3.3
requests==2.32.3
streamlit==1.37.0

